requirements.txt
# FastAPI and web server
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# LangChain and LLM
# Pin to 0.3.x for compatibility (1.0+ has breaking changes)
langchain>=0.3.0,<1.0.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-chroma>=0.2.0
langchain-huggingface>=0.0.1
langchain-google-genai>=2.0.0
langchain-text-splitters>=0.3.0

# Vector database
chromadb>=0.5.15

# Document processing
pypdf==4.3.1
tiktoken==0.7.0

# Embeddings
sentence-transformers==2.7.0

# Utilities
python-dotenv==1.0.1
matplotlib==3.9.2

